Creating test pipeline (#1)
This creates a `test` branch which builds a `test` Docker container. The
intention here is to be able to make changes to the bot, commit them to
the `test` branch to have it build a test container without impacting
the live container. I can then deploy the test container at the same
time as the live container to test changes made to the bot. Once the
changes are good and run smoothly from the test container, it may then
be merged into `main` where the production container will be built and
may be updated.