Creating test docker build on all non-main branches (#3)
After considering, being the only developer on this myself, I decided to
remove the whole "test" branch for a specific test flow. Instead, all
non-main branches will deploy to the test container. If it ever (very
VERY doubtful for this bot) ever gets to the point of multiple
developers working on this, adding separate Docker test flows, or having
a test branch would make sense as in that instance we would conflict
with each other when testing.

However, since it is JUST me doing development, I don't need to split
out the containers like that.