Prevent duplicate cards returned (#10)
This closes #8.

When duplicate card names are added such as `[[odin]] [[odin]]`, the
duplicates will be reduced to a single card.

Additionally, this restructures the CI/CD pipeline as it had an issue
before where the Docker container was being built using the `main`
branch for some reason. It was very odd, but produced a difficult
testing situation.