Better Search and Token Lookup (#11)
This revamps the way tokens are searched for. It might be slightly less
intuitive, but I feel like it provides an overall better user
experience.

When someone searches for a base card like `[[thanos]]`, this will check
to see what cards may be created by that card (e.g. Soul Stone, Time
Stone, etc.) and it will automatically pull those in, producing an
output that includes both the base card and the added tokens.

In addition to that, if someone searches directly for a token like
`[[Soul Stone]]`, this will detect that it is a token and will replace
the token with the base card (e.g. `[[Thanos]]`), then it will resolve
like normal, which will pull in all associated cards.