Adding formatting to summoned cards (#12)
Given the large amount of cards some cards (like Thanos and Nico) can
summon, it became a bit difficult to read and tell where one summon
ended and the next card began at a quick glance. This changes that by
adding a bullet-point list to all cards that are summoned by another
card, when the base card is listed in the response.